requirements.txt
argparse
numpy
scipy
matplotlib
torchvision
pillow
os